% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Generate a Greeting}
\usage{
hello_world(name = "World")
}
\arguments{
\item{name}{A character string. The name of the person or entity to greet.
Defaults to "World".}
}
\value{
A character string of the form "Hello, \link{name}!".
}
\description{
This function creates a standardized greeting string. It is useful for
demonstrating package structure and unit testing.
}
\examples{
hello_world()
hello_world("User")

\dontrun{
  # This will trigger an error due to invalid input type
  hello_world(123)
}
}
